refactored commands (unfinished)
